% by okada_n 120411
% charset=utf-8
\encoding{utf-8}
\name{adipocytesize.area}
\alias{adipocytesize.area}
\title{
[AdipocyteSize] Process Images
}
\description{
RImageJを用いて画像処理を行う。
}
\usage{
adipocytesize.area (dat, ver = "120508")
}
\arguments{
\item{dat}{使用するデータの入っているディレクトリ名。}
\item{ver}{使用する関数のバージョン。}
}
