% charset=utf-8
\encoding{utf-8}
\name{adipocytesize.histogram}
\alias{adipocytesize.histogram}
\title{
[AdipocyteSize] Plot Histgram
}
\description{
ヒストグラムを描画する。
}
\usage{
adipocytesize.histogram (dat, use, opt)
}
\arguments{
\item{dat}{使用するデータの日付。}
\item{scl}{１ピクセルの大きさ。単位はum（マイクロメートル）。}
\item{wht}{組織重量データの日付。}
\item{wid}{区切り幅。}
\item{sum}{各グループでの合計にするかどうか。}
}
