% by okada_n 120406
% charset=utf-8
\encoding{utf-8}
\name{adipocytesize.ready}
\alias{adipocytesize.ready}
\title{
[AdipocyteSize] Ready Images
}
\description{
RImageJを用いて画像処理を行う。人海戦術で細胞を特定するための画像を用意する。
}
\usage{
adipocytesize.ready (dat, ver = "120404")
}
\arguments{
\item{dat}{使用するデータの入っているディレクトリ名。}
\item{ver}{バージョン。}
}
