% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_plot.R
\name{ba_plot}
\alias{ba_plot}
\title{Bland-Altman ggplot2}
\usage{
ba_plot(df, x, y, xlabel = "Mean", ylabel = "Difference",
  measurement = "Measurement", ...)
}
\arguments{
\item{df}{Data frame}

\item{x}{Measurement 1}

\item{y}{Measurement 2}

\item{xlabel}{X-axis label (Difference between)}

\item{ylabel}{Y-axis label (Difference between)}

\item{measurement}{Description of what has been measured and is being compared}
}
\description{
Procduces customisable Bland-Altman Plots using ggplot2
}
\details{
Produces Bland-Altman plots of repeated measurements.  It first
standardises variables before calculating the mean and difference
of measurements for plotting.
}

