% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cache}
\alias{cache}
\title{Caches the result of long-running expressions to avoid unnecessary
re-execution.  The result of the expression is either cached and returned
or returned directly from cache.  If the source code within the expression
changes, then the result will be automatically re-calculated.}
\usage{
cache(expr, verbose = FALSE, clear_cache = FALSE, cache_dir = ".Rcache",
  cache_db = ".nmisc.cache")
}
\arguments{
\item{expr}{The expression whose result is cached.}

\item{verbose}{Is more verbose output needed?}

\item{clear_cache}{Should all of the cached objects be deleted?}

\item{cache_dir}{The directory containing the cached values.}

\item{cache_db}{The name of the stash/database used to cache the values.}
}
\value{
The result of the expression either from cache or freshly computed.
}
\description{
Caches the result of long-running expressions to avoid unnecessary
re-execution.  The result of the expression is either cached and returned
or returned directly from cache.  If the source code within the expression
changes, then the result will be automatically re-calculated.
}
\examples{
cache ({ 2 + 3 + 4 })
}

