% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams.R
\name{cams}
\alias{cams}
\title{Summarisation of Climbing Cams}
\usage{
cams(id = "", passwd = "", compare = "all", cam.doc = "Cam Size Data",
  cam.sheet = "Plain List (2015-12-19)", df = cams.df, smooth = "loess",
  free.scales = "free_y", wrap.col = 6, text.size = 16,
  exclude.outlier = FALSE, ...)
}
\arguments{
\item{id}{Google ID.}

\item{passwd}{Google Password.}

\item{cam.doc}{Name of Google Sheet holding data (don't change this).}

\item{cam.sheet}{Name of worksheet within Google Sheet holding data (don't change this).}

\item{smooth}{Smoothing function for plots.}

\item{free.scales}{Whether to allow the scales of faceted plots to be free.}

\item{wrap.col}{Number of columns to wrap \code{facet_wrap} by.}

\item{text.size}{Size of text in graphs in pts.}

\item{exclude.outlier}{Excludes the Valley Giant 12 from smoothing.}

\item{cams.df}{Name of R data frame version of data.}

\item{wrap.row}{Number of rows to wrap \code{facet_wrap} by.}
}
\description{
Read and summarise Climbing Cam data from Google Sheets
}

