% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{check_package_style}
\alias{check_package_style}
\title{Check R code styling for an entire package.}
\usage{
check_package_style(package_path = ".")
}
\arguments{
\item{package_path}{The path to the root directory of your package.}
}
\description{
Check adherence to a given style, syntax errors and possible
semantic issues for all R files.
}
\examples{
\donttest{check_package_style(package_path = ".")}
\donttest{check_package_style()}

}
\seealso{
Other code styling functions: \code{\link{check_style}},
  \code{\link{style_package}}, \code{\link{style_script}}
}
