% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{clear_warnings}
\alias{clear_warnings}
\title{Avoid R’s Warning Feature}
\usage{
clear_warnings()
}
\value{
Does not return anything.
}
\description{
Clears warnings for production code
}
\seealso{
\code{\link{warning}}
}
