% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-misc.R
\name{discard_at}
\alias{discard_at}
\title{Discard elements using their names}
\usage{
discard_at(.x, .at)
}
\arguments{
\item{.x}{a table of data}

\item{.at}{character (names) or a numeric (positions)}
}
\value{
a table of data
}
\description{
Discard elements using their names
}
