% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ecache}
\alias{ecache}
\title{Expression Cache}
\usage{
ecache(expr, key = NULL, clear_cache = getOption("clear_cache", default =
  FALSE), cache_dir = getOption("cache_dir", default = ".Recache"),
  verbose = getOption("verbose", default = FALSE))
}
\arguments{
\item{expr}{The expression to cache.}

\item{key}{[Optional] The key to cache the expression under.  In most cases, this
argument is not necessary.}

\item{clear_cache}{Should all cached objects be deleted?}

\item{cache_dir}{The directory containing the cache.  This should either be an
absolute path or relative to the current working directory.}

\item{verbose}{Is more verbose output needed?}
}
\value{
The expression's result either retrieved from cache or freshly calculated.
}
\description{
Caches the result of long-running expressions to avoid unnecessary
re-execution.  The function will return the cached result of the
expression if one exists.  Otherwise the function will execute the
expression, cache the result and then return it.
}
\details{
A checksum of an expression's source code, excluding white space,
is used to generate a unique identitifier to identify the cached expression.
The cache will be invalidated if the source code within the expression
itself changes.
}
\examples{
x <- ecache ({ 2 + 3 + 4 })
y <- ecache ({ Sys.sleep (2); 2 + 3 + 4})
}

