% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{fix_style}
\alias{fix_style}
\title{Fix R coding style
 
Fix R coding style issues in a specified file or directory, 
 based on Nmisc style}
\usage{
fix_style(path = getwd())
}
\arguments{
\item{path}{The path to the file or directory you want to check.}
}
\description{
Fix R coding style
 
Fix R coding style issues in a specified file or directory, 
 based on Nmisc style
}
\examples{
\donttest{fix_style("file_name.R")}

}
