% by matsubara 121017
% charset=utf-8
\encoding{utf-8}
\name{fluorescence.change}
\alias{fluorescence.change}
\title{
[Fluorescence] change Images
}
\description{
RImageJを用いて画像処理を行う。画像の右側を蛍光画像に取り替える。
}
\usage{
fluorescence.change (dat, ver = "121005")
}
\arguments{
\item{dat}{使用するデータの入っているディレクトリ名。}
\item{ver}{バージョン。}
}
