% by okada_n 120406
% charset=utf-8
\encoding{utf-8}
\name{fluorescence.copy}
\alias{fluorescence.copy}
\title{
[Fluorescence] Copy Images from Original, changing filename
}
\description{
画像ファイルをオリジナルからcopyして名前を書き換える。
}
\usage{
fluorescence.copy (dat)
}
\arguments{
\item{dat}{使用するdataの入っているdirectory名。}
}
