% by okada_n
% charset=utf-8
\encoding{utf-8}
\name{fluorescence.histogram}
\alias{fluorescence.histogram}
\title{
[Fluorescence] Plot Histgram
}
\description{
ヒストグラムを描画する。
}
\usage{
fluorescence.histogram (dat, wid = 10, scl = 2.78, sum = TRUE, plot = TRUE, ver = "120718")
}
\arguments{
\item{dat}{使用するデータの日付。}
\item{scl}{１ピクセルの大きさ。単位はum（マイクロメートル）。}
\item{wid}{区切り幅。}
\item{sum}{各グループでの合計にするかどうか。}
\item{plot}{描画するかどうか。}
\item{ver}{使用する関数のバージョン。}
}
