% by okada_n 120406
% charset=utf-8
\encoding{utf-8}
\name{fluorescence.ready}
\alias{fluorescence.ready}
\title{
[Fluorescence] Ready Images
}
\description{
RImageJを用いて画像処理を行う。人海戦術で細胞を特定するための画像を用意する。
}
\usage{
fluorescence.ready (dat, ver = "120628")
}
\arguments{
\item{dat}{使用するdataの入っているdirectory名。}
\item{ver}{version}
}
