% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-time.R
\name{format_utc}
\alias{format_utc}
\title{Format Date and POSIXct}
\usage{
format_utc(x, format = NULL, usetz = TRUE)
}
\arguments{
\item{x}{An object to be converted.}

\item{format}{A character string. The default for the  format methods is 
\"\%Y-\%m-\%d \%H:\%M:\%S\" for POSIXct and \"\%Y-\%m-\%d\" for Date.}

\item{usetz}{logical. Should the time zone abbreviation 
be appended to the output?}
}
\value{
character vectors representing the time.
}
\description{
Converts Date and POSIXct objects to a given format.
}
\seealso{
\code{\link{format.POSIXct}}
}
