% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{get_os}
\alias{get_os}
\title{Returns the name of the Operating Sytem}
\usage{
get_os()
}
\value{
One of \code{"win"}, \code{"mac"}, \code{"unix"}, 
 \code{"Unknown OS"}.
}
\description{
A simple wrapper arround \code{rappdirs:::get_os},
 allowing it to be exported.
}
