% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_if.R
\name{\%if_not_in\%}
\alias{\%if_not_in\%}
\title{Discards The Elements Present In A Set And Not Present In The Other}
\usage{
x \%if_not_in\% y
}
\description{
Unlike setdiff does not remove duplicates in x and keeps its order
}
