% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-time.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{Is it a POSIXct object?}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
Is it a POSIXct object?
}
\seealso{
\code{\link[lubridate:is.POSIXt]{lubridate::is.POSIXct}}
}
