% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-time.R
\name{is_POSIXct}
\alias{is_POSIXct}
\title{Date / Time functions}
\usage{
is_POSIXct(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
TRUE if the object is of class POSIXct
and FALSE otherwise.
}
\description{
Checks if the input of the function is a POXIXct object.
}
