% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_if.R
\name{keep_if_not_in}
\alias{keep_if_not_in}
\title{Discards The Elements Present In A Set And Not Present In The Other}
\usage{
keep_if_not_in(x, y)
}
\arguments{
\item{x, y}{Vectors (of the same mode) containing a sequence of items}
}
\value{
A vector of the same mode as x or y
}
\description{
Unlike setdiff does not remove duplicates in x and keeps its order
}
\examples{
keep_if_not_in(1:5, 3:6)
}
