% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpg.R
\name{mpg}
\alias{mpg}
\title{Summarisation of Fuel Efficiency}
\usage{
mpg(id = "", passwd = "", mpg.doc = "Fuel Efficiency (Responses)",
  mpg.sheet = "Form Responses 1", df = fuel, zoneinfo = "GMT", ...)
}
\arguments{
\item{id}{Your Google ID.}

\item{passwd}{Your Google Password.}

\item{mpg.doc}{Name of your Google Sheet holding data.}

\item{mpg.sheet}{Name of worksheet within Google Sheet holding data.}

\item{zoneinfo}{Your timezone.}
}
\description{
Read and summarise Fuel Efficiency Data from Google Sheets
}

