\name{nmisc-package}
\alias{nmisc-package}
\alias{nmisc}
\docType{package}
\title{
Miscellaneous functions useful to Nick.
}
\description{
Miscellaneous functions useful to Nick.
}
\details{
\tabular{ll}{
Package: \tab nmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-10\cr
License: \tab MIT\cr
}

}
\author{
Nick Allen <nick@nickallen.org>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
x <- ecache ({1 + 2 + 3})
}
