% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmisc_example.R
\name{nmisc_example}
\alias{nmisc_example}
\title{Runs Shiny for Cam comparison}
\usage{
nmisc_example(example)
}
\arguments{
\item{example}{The Shiny example you wish to run (omit for a list of possible options).}
}
\description{
Runs Shiny for Cam comparison
}

