% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-time.R
\name{now_utc}
\alias{now_utc}
\title{Current time in UTC time zone}
\usage{
now_utc(length = 1L)
}
\arguments{
\item{length}{Positive integer (scalar) indicating the length of the vector.}
}
\value{
A POSIXct vector of size \code{length} with the \code{tzone}
  atribute set to "UTC".
}
\description{
Returns a vector with the current date and time in the UTC time zone.
}
\examples{
now_utc(0)
# returns "POSIXct of length 0"

}
\seealso{
\code{\link{Sys.time}}, \code{\link[lubridate:now]{lubridate::now}}
}
