% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-misc.R
\name{pull_with_names}
\alias{pull_with_names}
\title{Pull out a single column}
\usage{
pull_with_names(.data, var = -1, name_col)
}
\arguments{
\item{.data}{A table of data}

\item{var}{The name of the column of interest, 
or a positive integer, giving the position counting from the left, 
or a negative integer, giving the position counting from the right.}

\item{name_col}{The column whose values will be used to
name the pulled column.}
}
\value{
A named vector.
}
\description{
Pull out a single column by using its name or its position
  and name the obtained vector using values from another column.
}
\examples{
pull_with_names(iris, 4, "Species")

}
