% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{seq_ncol}
\alias{seq_ncol}
\title{Creates A Sequence Based On Data Frame Number Of Columns}
\usage{
seq_ncol(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a vector of integers
}
\description{
Creates a sequence that starts at 1 and with steps of 1
which finishes at the number value corresponding to the
number of columns in a data frame given as input.
}
\seealso{
\code{\link{seq}}

\code{\link{seq_along}}
}
