% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq_nrow}
\alias{seq_nrow}
\alias{seq_ncol}
\title{Creates a sequence based on the number of rows or columns}
\usage{
seq_nrow(x)

seq_ncol(x)
}
\arguments{
\item{x}{a data frame or a matrix}
}
\value{
a vector of integers
}
\description{
Creates a sequence from 1 to the number of row or columns, respectively.
}
\seealso{
\code{\link{seq}}
}
