% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_dump.R
\name{slot_in}
\alias{slot_in}
\title{Insert a value into a vector after specified value}
\usage{
slot_in(vec, pos, val)
}
\arguments{
\item{vec}{Vector in which you insert the value}

\item{pos}{Element that you slot in after}

\item{val}{Element that you slot in}
}
\description{
Take a vector and put a value into it in a specified place
}
\examples{
slot_in()
}
\keyword{vector}

