% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Splits up a string by a given character}
\usage{
str_split_one(x, split)
}
\arguments{
\item{x}{str: string to be split up}

\item{split}{str: the string value to split x by}
}
\value{
array of characters
}
\description{
Splits up a string by a given character
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_one(x, split = ",")
}
