% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_package}
\alias{style_package}
\title{Perform R code styling on an R package.}
\usage{
style_package(package_path = ".")
}
\arguments{
\item{package_path}{The path of the root directory of the package
you want to style.}
}
\value{
Invisibly returns a data frame that indicates for
each file considered for styling whether or not it was
actually changed.
}
\description{
Pretty-print all R files without changing the user's formatting intent.
}
\section{Warning}{

This function overwrites files
(if styling results in a change of the code to be formatted).
It is strongly suggested to only style files that are under
version control or to create a backup copy.
}

\examples{
\donttest{style_package(package_path = ".")}
\donttest{style_package()}

}
\seealso{
Other code styling functions: \code{\link{check_package_style}},
  \code{\link{check_style}}, \code{\link{style_script}}
}
