% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_dump.R
\name{superNumeric}
\alias{superNumeric}
\title{Insert a value into a vector after specified value}
\usage{
superNumeric(vec)
}
\arguments{
\item{vec}{Vector in which you insert the value}
}
\description{
Normal numeric stuff is bad. This allows you to deal with the annoying characters that cause errors in normal is.numeric-ing
}
\examples{
superNumeric("$200,000.00")
}
\keyword{numeric}
\keyword{vector}

