% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_dump.R
\name{theme_nothing}
\alias{theme_nothing}
\title{One of many ggplot themes that make illustrator easier to use later.}
\usage{
theme_nothing(base_size = 12, base_family = "Optima")
}
\description{
ggplot needs to be stopped!
}
\examples{
ggplot(...) + theme_nothing()
}
\keyword{ggplot2}
\keyword{plotting}

