% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_dump.R
\name{theme_nyt_vert_horiz}
\alias{theme_nyt_vert_horiz}
\title{One of many ggplot themes that make illustrator easier to use later.}
\usage{
theme_nyt_vert_horiz(base_size = 12, base_family = "Helvetica")
}
\description{
ggplot needs to be stopped!
}
\examples{
ggplot(...) + theme_nyt_vert_horiz()
}
\keyword{ggplot2}
\keyword{plotting}

