% by okada_n 111114
% charset=utf-8
\encoding{utf-8}
\name{tree.aln}
\alias{tree.aln}
\title{
Alignment for Tree
}
\description{
系統樹を作成するためにアライメントを行う。
}
\usage{
tree.aln (name, file, tool = c("muscle", "clustalw"))
}
\arguments{
\item{name}{UNIPROT IDとACCの列を持つデータフレーム。}
\item{file}{アライメント結果を出力するファイル名。}
\item{program}{使用するプログラム。}
}
