% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{catn}
\alias{catn}
\title{Concatenate with Newline}
\usage{
catn(...)
}
\arguments{
\item{...}{arguments to be passed to cat function}
}
\value{
None
}
\description{
Wrapper to cat which appends new line to output
}
\seealso{
\code{\link{cat}}
}
