% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{catn}
\alias{catn}
\title{Concatenate with new line}
\usage{
catn(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[base]{cat}} function.}
}
\value{
None
}
\description{
Wrapper around cat which appends new line to output.
}
\seealso{
\code{\link{cat}}
}
