% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{check_style}
\alias{check_style}
\title{Check R code styling}
\usage{
check_style(file_name)
}
\arguments{
\item{file_name}{The name of the file you want to check.}
}
\description{
Check adherence to a given style, syntax errors and possible
  semantic issues.
}
\examples{
\donttest{check_style("file_name.R")}

}
