% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ecache}
\alias{ecache}
\title{Expression Cache}
\usage{
ecache(expr, verbose = FALSE, clear_cache = FALSE, cache_dir = ".Recache")
}
\arguments{
\item{expr}{The expression to cache.}

\item{verbose}{Is more verbose output needed?}

\item{clear_cache}{Should all cached objects be deleted?}

\item{cache_dir}{The directory containing the cache.  This should either be an
absolute path or relative to the current working directory.}
}
\value{
The expression's result either retrieved from cache or freshly calculated.
}
\description{
Caches the result of long-running expressions to avoid unnecessary
re-execution.  The function will return the cached result of the
expression if one exists.  Otherwise the function will execute the
expression, cache the result and then return it.
}
\details{
A checksum of an expression's source code, excluding white space,
is used to generate a unique identitifier to identify the cached expression.
The cache will be invalidated if the source code within the expression
itself changes.
}
\examples{
ecache ({ 2 + 3 + 4 })
ecache ({ Sys.sleep (2); 2 + 3 + 4})
}

