% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_packages.R
\name{generate_install_file}
\alias{generate_install_file}
\title{Takes a data frame containing package information and generates an install
file based on it}
\usage{
generate_install_file(packages_df, include_core_packages = FALSE)
}
\arguments{
\item{packages_df}{A data frame obtained with \code{get_packages} that 
contains information regarding the name, version and source of the package.}

\item{include_core_packages}{Logical, whether or not to include in the 
generated install file packages which come with R by default}
}
\value{
Nothing
}
\description{
the function takes the output of \code{get_packages} and
 writes in a file the commands needed in order to install
 packages used through the project that are not already installed.
}
\examples{
packages_df <- get_packages(
               'included_file\\\\.R(md)?$', 
               'excluded_file\\\\.R(md)?$', 
                c('required', 'library'))
generate_install_file(packages_df)

}
\seealso{
\code{\link{get_packages}}
}
