% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_packages.R
\name{get_packages}
\alias{get_packages}
\title{Get information about the packages used in the project}
\usage{
get_packages(include_pattern = "\\\\.R(md)?$", exclude_pattern = "tests|^_",
  package_options = c("library", "required", "referenced", "description"))
}
\arguments{
\item{include_pattern}{A string representing a regex that matches 
project files in which to look for packages. By default, \code{get_packages}
includes all .R files in the current project}

\item{exclude_pattern}{A string representing a regex that matches 
project files in which not to look for packages. By default, 
\code{get_packages} excludes all files found in "tests" folder.}

\item{package_options}{A character vector that represents the method through
which packages are loaded or referenced. The options are: \code{libray} for 
packages loaded using  \code{libray()}, \code{required} for 
packages loaded using  \code{require()}, \code{referenced} for 
packages referenced using  \code{::} operator, \code{loaded}, used to 
include packages already loaded in the current session and \code{description}
used to include packages mentioned in \code{DESCRIPTION} file}
}
\value{
A data frame containing package information.
}
\description{
the function returns a data frame containing information regarding 
 packages that are loaded with \code{library()}, \code{require()}, used
 with \code{::} operator and/or already loaded packages
}
\examples{
get_packages(
'\\\\.R(md)?$', 
'test\\\\.R(md)?$', 
c('required', 'library'))

}
