% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-misc.R
\name{keep_if_in}
\alias{keep_if_in}
\title{Performs Set Intersection}
\usage{
keep_if_in(x, y)
}
\arguments{
\item{x, y}{Vectors (of the same mode) containing a sequence of items}
}
\value{
A vector of the same mode as x or y
}
\description{
Unlike intersect does not remove duplicates in x and keeps its order
}
\examples{
keep_if_in(1:5, 3:6)

}
