% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-misc.R
\name{pull_with_names}
\alias{pull_with_names}
\title{Pull out a single column using another column for names}
\usage{
pull_with_names(.data, var = -1, name_col)
}
\description{
First verision, var must be character or numeric
}
