% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq_nrow}
\alias{seq_nrow}
\title{Creates A Sequence Based On Data Frame Number Of Rows}
\usage{
seq_nrow(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a vector of integers
}
\description{
Creates a sequence that starts at 1 and with steps of 1
which finishes at the number value corresponding to the
number of rows in a data frame given as input.
}
\seealso{
\code{\link{seq}}

\code{\link{seq_len}}
}
