% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str1}
\alias{str1}
\alias{str2}
\title{High level overview of the structure of an R object}
\usage{
str1(x)

str2(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Does not return anything.
}
\description{
\code{str1()} is a wrapper around \code{\link{str}} which sets maximal level
  of nesting to 1, while \code{str2()} sets maximal level of nesting to 2.
}
\seealso{
\code{\link{str}}
}
