% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str1}
\alias{str1}
\title{Compactly Display The Structure Of An Arbitrary R Object}
\usage{
str1(x)
}
\arguments{
\item{x}{any R object about which you want to have some information}
}
\value{
Does not return anything.
The obvious side effect is output to the terminal.
}
\description{
Wrapper to str which sets maximal level of nesting to 1.
}
\seealso{
\code{\link{str}}
}
